#6 ディレクトリ構成を整理
