--- app.yaml
@@ -1,1 +1,0 @@
-runtime: go112
